% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_R.R
\name{loglik_R}
\alias{loglik_R}
\title{R-based log-likelihood implementation}
\usage{
loglik_R(theta, W, X, Y, weights, nReg, ...)
}
\description{
Only kept as an internal reference; for pedagogical purposes...
}
