% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_frame.R
\name{model.frame.opsr}
\alias{model.frame.opsr}
\title{Extracting the Model Frame from OPSR Model Fits}
\usage{
\method{model.frame}{opsr}(formula, ...)
}
\arguments{
\item{formula}{an object of class \code{"opsr"}.}

\item{...}{a mix of further arguments such as \code{data}, \code{na.action} or \code{subset},
passed to the default method.}
}
\value{
A \code{\link{data.frame}} containing the variables used in \code{formula$formula}.
}
\description{
Extracting the Model Frame from OPSR Model Fits
}
\seealso{
\code{\link[stats:model.frame]{stats::model.frame}}
}
