% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{package}
\name{OPSR-package}
\alias{OPSR}
\alias{OPSR-package}
\title{Fitting ordinal probit switching regression (OPSR) models}
\description{
Some description of what the package does and pointing to resources/references.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/dheimgartner/OPSR/issues}
}

}
\author{
\strong{Maintainer}: Daniel Heimgartner \email{d.heimgartners@gmail.com} (\href{https://orcid.org/0000-0002-0643-8690}{ORCID})

Authors:
\itemize{
  \item Xinyi Wang \email{email@gmail.com} (\href{https://orcid.org/1234}{ORCID})
}

}
