% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsr_fit.R
\name{opsr.fit}
\alias{opsr.fit}
\title{Fitter function for ordinal probit switching regression model}
\usage{
opsr.fit(
  Ws,
  Xs,
  Ys,
  start,
  fixed,
  weights,
  method,
  iterlim,
  printLevel,
  nThreads,
  .useR = FALSE,
  ...
)
}
\arguments{
\item{weights}{a vector of weights to be used in the fitting process. Has to
conform with order (\code{w <- weights[order(Z)]}, where Z is the ordinal
outcome).}

\item{...}{}
}
\value{
object of class \verb{"maxLik" "maxim"}.
}
\description{
This is the basic computing engine called by \code{\link{opsr}} used to fit ordinal
probit switching regression models. Should usually \emph{not} be used directly.
Log-likelihood function is implemented in C++ which yields a considerable
speed-up. However, gradient and Hessian is currently numerically determined.
}
\seealso{
\link{'loglik'}, Formula 6 (translates almost verbatim).
}
