% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsr_2step.R
\name{opsr_2step}
\alias{opsr_2step}
\title{Heckman two-step estimation}
\usage{
opsr_2step(W, Xs, Z, Ys)
}
\description{
Tow-step estimation procedure to generate reasonable starting values.
}
\section{Note}{

Since the Heckman two-step estimator includes an estimate in the second step
regression, the resulting OLS standard errors and heteroskedasticity-robust
standard errors are incorrect.
}

