% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsr_generate_start.R
\name{opsr_generate_start}
\alias{opsr_generate_start}
\title{2-step regression to generate reasonable starting values}
\usage{
opsr_generate_start(W, Xs, Z, Ys)
}
\arguments{
\item{Ys}{}
}
\description{
Currently, separate ordinal probit regression and linear regression and
setting sigma and rho to 0.5. Should be improved!
}
\examples{
\dontrun{
## tick: this is done by opsr() internally
sim_dat <- opsr_simulate()
dat <- sim_dat$data
W <- as.matrix(dat[, c("xs1", "xs2")])
Z <- dat$ys
Y <- dat$yo
nReg <- length(unique(Z))
Xs <- lapply(seq_len(nReg), function(i) {
  X <- W
  X[Z == i, ]
})
Ys <- lapply(seq_len(nReg), function(i) {
  Y[Z == i]
})
## tock

opsr_generate_start(W, Xs, Z, Ys)
}
}
