% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{opsr_max_threads}
\alias{opsr_max_threads}
\title{Check Maximum Number of Threads Available}
\usage{
opsr_max_threads()
}
\value{
integer
}
\description{
Check Maximum Number of Threads Available
}
\seealso{
\code{\link{opsr_check_omp}}
}
