% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsr_model_matrices.R
\name{opsr_model_matrices}
\alias{opsr_model_matrices}
\title{Construct design matrices}
\usage{
opsr_model_matrices(object, yo, newdata = NULL, z = NULL)
}
\arguments{
\item{object}{an object of class \code{"opsr"}.}

\item{yo}{creates the design matrix for the process of group \code{yo} (ordered
outcome level).}

\item{newdata}{an optional data frame in which to look for varialbes used in
\code{object$formula}. If omitted, \code{environment(object$formula)} is used.}

\item{z}{Sometimes it is useful to create the design matrix for \code{yo} but filter
the observations belonging to another group \code{z} (e.g. in \code{\link{predict.opsr}}).}
}
\value{
list of design matrices \code{X_j} and \code{W_j}.
}
\description{
Somewhat similar to \code{\link{model.matrix}} creates a design (or model) matrix from
the \code{formula} (attached to \code{object}) used when fitting the OPSR model. Creates
the design matrix for a specific outcome group \code{yo}.
}
