% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsr_null_model.R
\name{opsr_null_model}
\alias{opsr_null_model}
\title{Null Model for OPSR Model fits}
\usage{
opsr_null_model(object, ...)
}
\arguments{
\item{object}{an object of class \code{"opsr"}.}

\item{...}{further arguments passed to \code{\link{opsr}}.}
}
\value{
An object of class \verb{"opsr.null" "opsr"}.
}
\description{
Intercept-only model with no error correlation.
}
\examples{
sim_dat <- opsr_simulate()
dat <- sim_dat$data
model <- ys | yo ~ xs1 + xs2 | xo1 + xo2
fit <- opsr(model, dat)
fit_null <- opsr_null_model(fit)
summary(fit_null)
}
