% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsr_simulate.R
\name{opsr_simulate}
\alias{opsr_simulate}
\title{Simulate Data from an OPSR Process}
\usage{
opsr_simulate(nobs = 1000, sigma = NULL)
}
\arguments{
\item{nobs}{number of observations to simulate.}

\item{sigma}{the covariance matrix of the multivariate normal.}
}
\value{
Named list.
}
\description{
Simulates data from an ordinal probit process and separate (for each regime)
OLS process where the errors follow a multivariate normal distribution.
}
\details{
The named list contains the ground truth parameters \code{params}, the simulated
\code{data}, the error draws \code{errors} and the assumed covariance matrix \code{sigma}.
Three ordinal outcomes are simulated and the distinct design matrices (\code{W} and
\code{X}) are used (if \code{W == X} the model is poorely identified). Variables \code{ys} and
\code{xs} in \code{data} correspond to the selection process and \code{yo}, \code{xo} to the outcome
process.
}
