% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_opsr.R
\name{plot.opsr}
\alias{plot.opsr}
\title{Plot Method for OPSR Model Fits}
\usage{
\method{plot}{opsr}(
  x,
  type = c("response", "unlog-response", "prob", "mills", "correction", "Xb"),
  weights = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"opsr"}.}

\item{type}{see \code{\link{predict.opsr}} for details.}

\item{weights}{a vector of weights. If \code{NULL} then weights from \code{object} will
be used.}

\item{...}{further arguments passed to \code{\link{pairs.opsr.te}}.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Wrapper around \code{\link{pairs.opsr.te}}.
}
\seealso{
\code{\link{opsr_te}}, \code{\link{pairs.opsr.te}}
}
