% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.opsr}
\alias{predict.opsr}
\title{Predict method for OPSR model fits}
\usage{
\method{predict}{opsr}(
  object,
  newdata,
  group,
  counterfact = NULL,
  type = c("response", "unlog-response", "prob", "mills"),
  ...
)
}
\arguments{
\item{object}{an object of class \code{"opsr"}.}

\item{newdata}{an optional data frame in which to look for variables used in
\code{object$formula}. If omitted, \code{environment(object$formula)} is used. See
also \code{\link{model.matrix.opsr}}.}

\item{group}{predict outcome of this group.}

\item{counterfact}{counterfactual group.}
}
\value{
a vector of length \code{nrow(newdata)} (or data used during estimation).
}
\description{
Computes the conditional expectation \eqn{E[y_j | Z = j]} or
counterfactual conditional expectation \eqn{E[y_{j'} | Z = j]}.
}
\details{
Elements are \code{NA_real_} if the \code{group} does not correspond to the observed
regime (selection outcome). This ensures consistent output length.
}
\section{Outlook}{

Could be extended with an argument \code{se.fit}, computing the standard errors
of the predictions with the delta method.
}

