% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_opsr_diagnostics.R
\name{print.opsr.diagnostics}
\alias{print.opsr.diagnostics}
\title{Print Method for OPSR Convergence Diagnostics}
\usage{
\method{print}{opsr.diagnostics}(
  x,
  digits = max(3L, getOption("digits") - 3L),
  print.gradient = FALSE,
  print.eigenvalues = FALSE,
  .print.title = TRUE,
  ...
)
}
\arguments{
\item{x}{an object of class \code{"opsr.diagnostics"}.}

\item{digits}{minimum number of significant digits for numeric output.}

\item{print.gradient}{if \code{TRUE}, prints the full gradient vector.}

\item{print.eigenvalues}{if \code{TRUE}, prints all eigenvalues of the Hessian.}

\item{.print.title}{flag to disable printing when called from \code{\link{print.summary.opsr}}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Prints diagnostics and returns \code{x} invisibly.
}
\description{
Print Method for OPSR Convergence Diagnostics
}
\seealso{
\code{\link{opsr_diagnostics}}
}
