% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_opsr_te.R
\name{print.opsr.te}
\alias{print.opsr.te}
\title{Print Method for OPSR ATE Objects}
\usage{
\method{print}{opsr.te}(x, ...)
}
\arguments{
\item{x}{an object of class \code{"opsr.te"}.}

\item{...}{further arguments passed to \code{\link{summary.opsr.te}}.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Print Method for OPSR ATE Objects
}
\details{
This is just a wrapper around \code{\link{summary.opsr.te}} and a subsequent call to
\code{\link{print.summary.opsr.te}}.
}
\seealso{
\code{\link{print.summary.opsr.te}}
}
