% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_opsr_te.R
\name{summary.opsr.te}
\alias{summary.opsr.te}
\title{Summarizing OPSR TE Objects}
\usage{
\method{summary}{opsr.te}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"opsr.te"}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
An object of class \code{"summary.opsr.te"} containing among others:
\itemize{
\item \code{ate}: An object of class \code{"ate"}.
\item \code{te}: An object of class \code{"te"}.
}

The p-values of the weighted paired t-test are attached as attributes.
}
\description{
This function computes weighted treatment effects and corresponding weighted
paired t-tests.
}
