% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{data}
\name{telework_data}
\alias{telework_data}
\title{Telework data}
\format{
Data frame with columns
\describe{
\item{variable1}{Description.}
\item{variable2}{Description.}
}
}
\source{
tbc
}
\usage{
telework_data
}
\description{
Telework data
}
\keyword{datasets}
