% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik_R.R
\name{loglik_R}
\alias{loglik_R}
\title{R-based Log-Likelihood Implementation}
\usage{
loglik_R(theta, W, X, Y, weights, nReg, ...)
}
\description{
R-based Log-Likelihood Implementation
}
\seealso{
\code{\link{loglik_cpp}}, \code{\link{opsr.fit}}
}
\keyword{internal}
