% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{package}
\name{OPSR-package}
\alias{OPSR}
\alias{OPSR-package}
\title{OPSR: Ordinal Probit Switching Regression}
\description{
Estimates ordinal probit switching regression models - a Heckman type selection model with an ordinal selection and continuous outcomes. Different model specifications are allowed for each treatment/regime.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dheimgartner/OPSR}
  \item Report bugs at \url{https://github.com/dheimgartner/OPSR/issues}
}

}
\author{
\strong{Maintainer}: Daniel Heimgartner \email{d.heimgartners@gmail.com} (\href{https://orcid.org/0000-0002-0643-8690}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Xinyi Wang \email{xinyi174@mit.edu} (\href{https://orcid.org/0000-0002-3564-9147}{ORCID})
}

}
