% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{package}
\name{OPSR}
\alias{OPSR}
\alias{OPSR-package}
\title{Fitting ordinal probit switching regression (OPSR) models}
\description{
Some description of what the package does and pointing to resources/references.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
