% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opsr_generate_start.R
\name{opsr_generate_start}
\alias{opsr_generate_start}
\title{2-step regression to generate reasonable starting values.}
\usage{
opsr_generate_start(W, Xs, Z, Ys)
}
\arguments{
\item{Ys}{}
}
\description{
Currently, separate ordinal probit regression and linear regression and
setting sigma to 1 and rho to 0. Should be improved!
}
\examples{
\dontrun{
## generate the arguments (as done by opsr())
sim_dat <- sim_dat_1()
dat <- sim_dat$data
W <- as.matrix(dat[, c("X1", "X2")])
Z <- dat$Z
Y <- dat$Y
nReg <- length(unique(Z))
Xs <- lapply(seq_len(nReg), function(i) {
  X <- W
  X[Z == i, ]
})
Ys <- lapply(seq_len(nReg), function(i) {
  Y[Z == i]
})

opsr_generate_start(W, Xs, Z, Ys)
}
}
