% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.opsr}
\alias{predict.opsr}
\title{Predict method for OPSR model fits}
\usage{
\method{predict}{opsr}(object, newdata, group, counterfact = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{"opsr"}.}

\item{newdata}{an optional data frame in which to look for variables used in
\code{object$formula}. If omitted, \code{environment(object$formula)} is used. See
also \code{\link{model.matrix.opsr}}.}

\item{group}{predict outcome of this group.}

\item{counterfact}{counterfactual group.}
}
\description{
Computes the conditional expectation \eqn{E[y_j | Z = j]} or
counterfactual conditional expectation \eqn{E[y_{j'} | Z = j]}.
}
\section{Outlook}{

Could be extended with an argument \code{se.fit}, computing the standard errors
of the predictions with the delta method.
}

